Factor decrypt module out of e2ee module
